fix: Implement CR suggestions
